ai-engine/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
scikit-learn==1.6.1
numpy==2.2.4
joblib==1.4.2
langchain==0.3.23
langchain-core==0.3.52
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5

